bansi

+++ base/com/controller/state_controller.py
@@ -1,15 +1,16 @@
 from flask import *
-
 from base import app
 from base.com.dao.state_dao import StateDAO
 from base.com.vo.state_vo import StateVO
-
+from base.com.controller.login_controller import admin_login_session
 
 @app.route('/admin/load_state')
 def admin_load_state():
     try:
-        return render_template('admin/addState.html')
-
+        if admin_login_session() == 'admin':
+            return render_template('admin/addState.html')
+        else:
+            return redirect(url_for('admin_logout_session'))
     except Exception as ex:
         print("admin_load_state route exception occured>>>>>>>>>>", ex)
 
@@ -17,15 +18,19 @@ def admin_load_state():
 @app.route('/admin/insert_state', methods=['post'])
 def admin_insert_state():
     try:
-        state_vo = StateVO()
-        state_dao = StateDAO()
+        if admin_login_session() == 'admin':
 
-        state_vo.state_name = request.form.get('stateName')
-        state_vo.state_description = request.form.get('stateDescription')
+            state_vo = StateVO()
+            state_dao = StateDAO()
 
-        state_dao.insert_state(state_vo)
+            state_vo.state_name = request.form.get('stateName')
+            state_vo.state_description = request.form.get('stateDescription')
 
-        return redirect(url_for('admin_view_state'))
+            state_dao.insert_state(state_vo)
+
+            return redirect(url_for('admin_view_state'))
+        else:
+            return redirect(url_for('admin_logout_session'))
 
     except Exception as ex:
         print("admin_insert_state route exception occured>>>>>>>>>>", ex)
@@ -34,9 +39,13 @@ def admin_insert_state():
 @app.route('/admin/view_state', methods=['get'])
 def admin_view_state():
     try:
-        state_dao = StateDAO()
-        state_vo_list = state_dao.view_state()
-        return render_template('admin/viewState.html', state_vo_list=state_vo_list)
+        if admin_login_session() == 'admin':
+
+            state_dao = StateDAO()
+            state_vo_list = state_dao.view_state()
+            return render_template('admin/viewState.html', state_vo_list=state_vo_list)
+        else:
+            return redirect(url_for('admin_logout_session'))
 
     except Exception as ex:
         print("admin_view_state route exception occured>>>>>>>>>>", ex)
@@ -45,11 +54,15 @@ def admin_view_state():
 @app.route('/admin/delete_state', methods=['get'])
 def admin_delete_state():
     try:
-        state_vo = StateVO()
-        state_dao = StateDAO()
-        state_vo.state_id = request.args.get('stateId')
-        state_dao.delete_state(state_vo)
-        return redirect(url_for('admin_view_state'))
+        if admin_login_session() == 'admin':
+
+            state_vo = StateVO()
+            state_dao = StateDAO()
+            state_vo.state_id = request.args.get('stateId')
+            state_dao.delete_state(state_vo)
+            return redirect(url_for('admin_view_state'))
+        else:
+            return redirect(url_for('admin_logout_session'))
 
     except Exception as ex:
         print("admin_delete_state route exception occured>>>>>>>>>>", ex)
@@ -58,11 +71,15 @@ def admin_delete_state():
 @app.route('/admin/edit_state', methods=['get'])
 def admin_edit_state():
     try:
-        state_vo = StateVO()
-        state_dao = StateDAO()
-        state_vo.state_id = request.args.get('stateId')
-        state_vo_list = state_dao.edit_state(state_vo)
-        return render_template('admin/editState.html', state_vo_list=state_vo_list)
+        if admin_login_session() == 'admin':
+
+            state_vo = StateVO()
+            state_dao = StateDAO()
+            state_vo.state_id = request.args.get('stateId')
+            state_vo_list = state_dao.edit_state(state_vo)
+            return render_template('admin/editState.html', state_vo_list=state_vo_list)
+        else:
+            return redirect(url_for('admin_logout_session'))
 
     except Exception as ex:
         print("admin_edit_state route exception occured>>>>>>>>>>", ex)
@@ -71,14 +88,18 @@ def admin_edit_state():
 @app.route('/admin/update_state', methods=['post'])
 def admin_update_state():
     try:
-        state_vo = StateVO()
-        state_dao = StateDAO()
-        state_vo.state_id = request.form.get('stateId')
-        state_vo.state_name = request.form.get('stateName')
-        state_vo.state_description = request.form.get('stateDescription')
+        if admin_login_session() == 'admin':
 
-        state_dao.update_state(state_vo)
-        return redirect(url_for('admin_view_state'))
+            state_vo = StateVO()
+            state_dao = StateDAO()
+            state_vo.state_id = request.form.get('stateId')
+            state_vo.state_name = request.form.get('stateName')
+            state_vo.state_description = request.form.get('stateDescription')
+
+            state_dao.update_state(state_vo)
+            return redirect(url_for('admin_view_state'))
+        else:
+            return redirect(url_for('admin_logout_session'))
 
     except Exception as ex:
         print("admin_update_state route exception occured>>>>>>>>>>", ex)
